feat: added a lot of improvements to AsyncTd views, and other things - created AsyncTdVideoPlayer - created AsyncTdFile, which now encapsulates all needed logic, and now backs the Image, QuickLookView, and VideoPlayer AsyncTd views - starting to implement messageVideo support - starting to move the album generation logic to it's respective function
