+++ Moc/Views/Chat/Message/MessageView+Video.swift
@@ -1,0 +1,15 @@
+//
+//  MessageView+Video.swift
+//  Moc
+//
+//  Created by Егор Яковенко on 24.06.2022.
+//
+
+import SwiftUI
+import TDLibKit
+
+extension MessageView {
+    func makeMessageVideo(from info: MessageVideo) -> some View {
+        EmptyView()
+    }
+}
